PPTX

Documentation for PPTX.

PPTX.PPTX

PPTX.Picture

PPTX.Presentation

PPTX.Slide

PPTX.TextBox

Base.write

PPTX.PPTX — Module

Interface functions:

Presentation: Presentation constructor, used for making a presentation.

Slide: Slide constructor, used for making a slide.

TextBox: TextBox constructor, used for adding text to slides.

Picture: Picture constructor, used for adding pictures to slides.

write: write a presentation to a file.

PPTX.Picture — Type

Picture(source::String; top::Int=0, left::Int=0, size::Int = 40)

source::String path of image file

top::Int mm from the top

left::Int mm from the left

Examples

julia> using PPTX

julia> img = Picture(joinpath(PPTX.EXAMPLE_DIR, "pictures/cauliflower.jpg"))

julia> slide = Slide()

julia> push!(slide, img)

julia> pres = Presentation()

julia> push!(pres, slide)

julia> write("cauliflower.pptx", pres)

PPTX.Presentation — Type

Presentation(
    slides::Vector{Slide}=Slide[];
    title::String="unknown",
    author::String="unknown",
)

Type to contain the final presentation you want to write to .pptx.

If isempty(slides) then we add a first slide with the Title slide layout.

Examples

julia> using PPTX

julia> pres = Presentation(; title = "My Presentation")

julia> slide = Slide()

julia> text = TextBox("Hello world!")

julia> push!(slide, text)

julia> push!(pres, slide)

julia> write("hello_world.pptx", pres)

PPTX.Slide — Type

Slide(
    shapes::Vector{AbstractShape}=AbstractShape[];
    title::String="",
    layout::Int=1,
)

shapes::Vector{AbstractShape} shapes to add to the PowerPoint, can also be pushed afterwards

title::String title text placed inside the title textbox found in the slide layout

layout::Int which slide layout to use. Typically 1 is the title slide and 2 is the text slide.

Make a Slide for a powerpoint Presentation.

You can push! any AbstractShape types into this slide, such as a TextBox or Picture.

Examples

julia> using PPTX

julia> slide = Slide(; title="Hello Title", layout=2)

julia> text = TextBox("Hello world!")

julia> push!(slide, text)

julia> pres = Presentation()

julia> push!(pres, slide)

julia> write("hello_world.pptx", pres)

PPTX.TextBox — Type

TextBox(;
    content::String = "",
    offset_x::Real = 50,
    offset_y::Real = 50,
    size_x::Real = 40,
    size_y::Real = 30,
    style::Dict = Dict("bold" => false, "italic" => false),
)

A TextBox to be used on a Slide. Offsets and sizes are in millimeters, but will be converted to EMU.

Examples

julia> using PPTX

julia> text = TextBox(content="Hello world!")

julia> slide = Slide()

julia> push!(slide, text)

julia> pres = Presentation()

julia> push!(pres, slide)

julia> write("hello_world.pptx", pres)

Base.write — Method

Base.write(
    filepath::String,
    p::Presentation;
    overwrite::Bool=false,
    open_ppt::Bool=true,
    template_path::String="no-slides.pptx",
)

filepath::String Desired presentation filepath.

pres::Presentation Presentation object to write.

overwrite = false argument for overwriting existing file.

open_ppt = true open powerpoint after it is written.

template_path::String path to an (empty) pptx that serves as template.

Examples

julia> using PPTX

julia> slide = Slide()

julia> text = TextBox("Hello world!")

julia> push!(slide, text)

julia> pres = Presentation()

julia> push!(pres, slide)

julia> write("hello_world.pptx", pres)